+++ CHANGELOG.md
@@ -5,6 +5,7 @@ dev:
   - relax proposal checks to enable DVT proposals
   - add individual "controller.fast-track" flags for attestations and sync committees
   - add 'failed' dimension for root to slot lookup metrics
+  - add 'deadline' builder bid strategy
 
 1.9.0-alpha.3
   - add proposal value and blinded status to trace